+++ CHANGELOG.md
@@ -9,6 +9,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 
+- Defend browser.stop() from non-responsive browser hangs or freezes @raycardillo
+
 ### Added
 
 ### Changed